Move permissions folder outside auth
